1.visitor添加字段 2.团购交易额刷新 3.卖家注册成功插入tdictionarys表数据 4.抽奖单下单时候判断人数

+++ src/main/java/com/haoyue/tuangou/pojo/TDictionarys.java
@@ -20,7 +20,7 @@ public class TDictionarys  {
     private String saleId;
     private int views;
     private int visitors;
-    private int turnover;
+    private double turnover;
 
     @JsonFormat(pattern="yyyy-MM-dd", timezone="GMT+8")
     @Temporal(TemporalType.DATE)
@@ -60,11 +60,11 @@ public class TDictionarys  {
         this.visitors = visitors;
     }
 
-    public int getTurnover() {
+    public double getTurnover() {
         return turnover;
     }
 
-    public void setTurnover(int turnover) {
+    public void setTurnover(double turnover) {
         this.turnover = turnover;
     }
